Set FD of IOBuf to -1 on close
This should prevent us from using the IOBuf to do anything useful after
a close.  I also added a convenience function for registering a
disconnect with an IOBuf.

Also, change the handling of an empty handler message to be in the
sending task, so that we don't close the FD until all data has been
flushed